+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Rule for Dutch postcodes
 - Rule for Dutch Social Security numbers
+- Rule for Dutch VAT numbers
 - Added Database/LessThanValue rule
 - Added Database/LessThanOrEqualValue rule
 - Added Database/MoreThanValue rule